+++ src/main/java/com/vnaskos/pdfguru/execution/util/PagePatternTranslator.java
@@ -1,4 +1,4 @@
-package com.vnaskos.pdfguru.execution;
+package com.vnaskos.pdfguru.execution.util;
 
 import java.util.ArrayList;
 import java.util.List;
