+++ src/test/java/com/vnaskos/pdfguru/execution/util/FileNamerTest.java
@@ -1,4 +1,4 @@
-package com.vnaskos.pdfguru.execution;
+package com.vnaskos.pdfguru.execution.util;
 
 import org.junit.Test;
